+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## Unreleased
+
+### Added
+
+- ORD Integration Dependencies for events consumed via `messaging.on()` are now declared using the `@OrdId` CDS annotation
+
 ## Version 0.3.1 - 2025-11-10
 
 ### Added